requirements.txt
appdirs==1.4.4
asgiref==3.3.1
autopep8==1.5.4
black==20.8b1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
cryptography==3.3.1
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.1.5
django-autoslug==1.9.8
django-baton==2.2.1
django-ckeditor==6.0.0
django-crispy-forms==1.10.0
django-extensions==3.1.0
django-filer==2.0.2
django-filter==2.4.0
django-js-asset==1.2.2
django-mptt==0.11.0
django-polymorphic==3.0.0
django-vendor==0.1.39
easy-thumbnails==2.7.1
gunicorn==20.1.0
idna==2.10
iso4217==1.6.20180829
isort==5.7.0
jdatetime==3.6.2
ldif3==3.1.1
mypy-extensions==0.4.3
oauthlib==3.1.0
pathspec==0.8.1
Pillow==8.1.0
pycodestyle==2.6.0
pycparser==2.20
PyJWT==2.0.1
python-card-me==0.9.3
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.5
regex==2020.11.13
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
surlex==0.2.0
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
unicodecsv==0.14.1
Unidecode==1.1.2
urllib3==1.26.2
whitenoise==5.2.0
